requirements.txt
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
fastapi==0.120.4
fonttools==4.60.1
gitdb==4.0.12
GitPython==3.1.45
guidata==3.13.1
h11==0.16.0
h5py==3.15.1
idna==3.11
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.7
narwhals==2.10.1
networkx==3.5
numpy==2.3.4
nvidia-nccl-cu12==2.28.7
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.3.1
PlotPy==2.8.0
protobuf==6.33.0
pyarrow==21.0.0
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PythonQwt==0.15.0
pytz==2025.2
QtPy==2.4.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.49.1
streamlit==1.51.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.10.16
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
xgboost==3.1.1
